requirements.txt
streamlit
torch
transformers
openai-whisper
